Initial implementation of higher-order constructs via 'macro'-style function synthesis.
This adds a 'macro-expansion' stage which is performed before
type-inference, which enables the synthesized code to be type-checked.
As a result, some checks (such as function argument arity checks and
the prevention of binding higher-order constructs) are now performed
in this stage instead of the type-checking phase, since they are
needed for function synthesis.

This stage is similar to the format-decoration stage; the difference
is that format decoration addresses the grammar sublanguage, while
this addresses the functional sublanguage.

A drawback is that errors in synthesized and replaced code might
result in confusing error messages from later stages of the pipeline.
It may be possible to avoid this by type-checking and analyzing
un-macro-expanded code, and then re-doing it on macro-expanded code
only if there were no errors.