Partial (syntax + type-checking) implementation of multi-assignments for map-views.
An example of the syntax is:

  @#[views, NT<i <- (m), j = p, k <- (n), l = o>]

which allows the attributes `i` and `k` to take values from lists
`m` and `n` as `NT` is applied to views `v` from the list `views`,
while attributes `j` and `l` take fixed values `p` and `o` for
each view `v` in `views`.

Code generation remains, as also the introduction of a constraint
to ensure that lists `m` and `n` have the same length as `views`.