Remove the failcont stack from the IR and the interpreter.
It was not really serving any purpose: when in the correct state,
it was merely used for asserting that the failcont being branched
to was the one on top of the stack.  It was sometimes not in the
correct state due to the need for cleanup trampolines in the failure
path, making correct stack management error-prone for little benefit.

Recover some of the static benefit by introducing an N_fail IR node,
which is functionally like N_jump except that it signifies a jump in
an error path.

Add some tests for whitespace decorators which uncovered the issue.