Make the typing of module values more uniform.
This removes the special case for typing `E_mod_member` values
in the 'local' module.  This requires macro-synthesized functions
to be specified using the `recfuns` AST node, and their calls to
be module-qualified, and hence requiring them to be exposed in
the module components in the IR.  We might want to revisit this
tradeoff, or use a different approach to address hiding synthesized
values in the IR.

This also removes the need for the typing environment to track the
current module.

For better error reporting, keep the typing environment up-to-date
with in-scope value definitions.